solo aparece borrar si no tiene actividad o venta

+++ resources/views/pages/prospectos_clientes.blade.php
@@ -91,7 +91,9 @@
          <td nowrap>
             <a href="/prospectos/{{ $prospecto->id }}"><i class="far fa-eye"></i></a>&nbsp;
             <a href="/prospectos/{{ $prospecto->id }}/form"><i class="far fa-edit"></i></a>&nbsp;
-            <a href="/prospectos/delete/{{ $prospecto->id }}"><i class="far fa-trash-alt"></i></a>
+            @if( $prospecto->hijos < 1  )
+              <a onclick="return confirm('¿Estas seguro de querer eliminar este prospecto?')" href="/prospectos/delete/{{ $prospecto->id }}"><i class="far fa-trash-alt"></i></a>
+            @endif
         </td>
         <td nowrap><i style="font-size:10px; color:{{ $prospecto->semaforo }}" class="fas fa-circle"></i> {{ $prospecto->empresa }} </td>          
         <td nowrap>{{ $prospecto->contacto }}</td>
